requirements.txt
flask
requests
twilio
openai
python-dotenv
